chore: move event models into separate file
